+++ js/core/utils/selection_filter.js
