+++ js/ui/selection/selection.strategy.standard.js
@@ -5,7 +5,7 @@ var commonUtils = require("../../core/utils/common"),
     getKeyHash = commonUtils.getKeyHash,
     dataQuery = require("../../data/query"),
     deferredUtils = require("../../core/utils/deferred"),
-    SelectionFilterCreator = require("../../core/utils/filter_creator").SelectionFilterCreator,
+    SelectionFilterCreator = require("../../core/utils/selection_filter").SelectionFilterCreator,
     when = deferredUtils.when,
     Deferred = deferredUtils.Deferred,
     errors = require("../widget/ui.errors"),
